Extract common methods to base class
